fix(tests): async issues in test_run_int.py — remove asyncio.run, add await, fix mocks
Fixes 12 failing tests in test_run_int.py after async migration:
- Remove asyncio.run(await ...) in async test functions; use await directly
- Update imports: async_runner → runner, async_latest_run → latest_run
- Make test methods async when calling async functions without AsyncMock
- Update test patches to reference non-async function names (IndexService not AsyncIndexService)
- Fix mock iterator for iter_raw_conversations to return proper async generator
- Use AsyncMock for patching async methods (_process_raw_batch)
