+++ polylogue/pipeline/services/parsing.py
@@ -1,6 +1,6 @@
-"""Parsing service for pipeline operations.
+"""Async parsing service for pipeline operations.
 
-This service implements the PARSE stage of the pipeline:
+This service implements the async PARSE stage of the pipeline:
 - Reads from raw_conversations (DB)
 - Parses into typed conversation/message records
 - Stores in the conversations/messages tables
@@ -10,14 +10,13 @@ It does NOT handle raw storage - that's the acquisition stage's job.
 
 from __future__ import annotations
 
-import concurrent.futures
+import asyncio
 import json
-import threading
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, TypeAlias
 
 from polylogue.lib.log import get_logger
-from polylogue.pipeline.prepare import prepare_records
+from polylogue.pipeline.async_prepare import async_prepare_records
 from polylogue.sources.source import _parse_json_payload
 from polylogue.storage.search_cache import invalidate_search_cache
 from polylogue.storage.store import RawConversationRecord
@@ -25,15 +24,15 @@ from polylogue.storage.store import RawConversationRecord
 if TYPE_CHECKING:
     from polylogue.config import Config, Source
     from polylogue.sources.parsers.base import ParsedConversation
-    from polylogue.storage.repository import ConversationRepository
+    from polylogue.storage.async_repository import AsyncConversationRepository
 
 logger = get_logger(__name__)
 
 _ProcessResult: TypeAlias = tuple[str, dict[str, int], bool]
 
 
-class ParseResult:
-    """Result of a parsing operation."""
+class AsyncParseResult:
+    """Result of an async parsing operation."""
 
     def __init__(self) -> None:
         self.counts: dict[str, int] = {
@@ -52,9 +51,9 @@ class ParseResult:
         self.processed_ids: set[str] = set()
         self.parse_failures: int = 0
         self.drift_counts: dict[str, int] = {}  # provider -> drift warning count
-        self._lock = threading.Lock()
+        self._lock = asyncio.Lock()
 
-    def merge_result(
+    async def merge_result(
         self,
         conversation_id: str,
         result_counts: dict[str, int],
@@ -64,12 +63,12 @@ class ParseResult:
 
         Args:
             conversation_id: ID of the processed conversation
-            result_counts: Count dictionary from prepare_records
+            result_counts: Count dictionary from async_prepare_records
             content_changed: Whether content hash changed
         """
         ingest_changed = (result_counts["conversations"] + result_counts["messages"] + result_counts["attachments"]) > 0
 
-        with self._lock:
+        async with self._lock:
             if ingest_changed or content_changed:
                 self.processed_ids.add(conversation_id)
             if content_changed:
@@ -83,24 +82,24 @@ class ParseResult:
                     self.counts[key] += value
 
 
-class ParsingService:
-    """Service for parsing conversations from sources.
+class AsyncParsingService:
+    """Service for parsing conversations from sources asynchronously.
 
-    This service implements the PARSE stage of the pipeline.
+    This service implements the async PARSE stage of the pipeline.
     It does NOT handle raw storage - use AcquisitionService for that.
     """
 
     def __init__(
         self,
-        repository: ConversationRepository,
+        repository: AsyncConversationRepository,
         archive_root: Path,
         config: Config,
         drive_client_factory: Any | None = None,
     ):
-        """Initialize the parsing service.
+        """Initialize the async parsing service.
 
         Args:
-            repository: Storage repository for database operations
+            repository: Async storage repository for database operations
             archive_root: Root directory for archived conversations
             config: Application configuration
             drive_client_factory: Optional factory callable returning a DriveClient
@@ -111,16 +110,16 @@ class ParsingService:
         self.drive_client_factory = drive_client_factory
         # Per-run drift tracking (reset each parse_from_raw call)
         self._drift_counts: dict[str, int] = {}
-        self._drift_lock = threading.Lock()
+        self._drift_lock = asyncio.Lock()
 
-    def parse_sources(
+    async def parse_sources(
         self,
         sources: list[Source],
         *,
         ui: object | None = None,
         download_assets: bool = True,
         progress_callback: Any | None = None,
-    ) -> ParseResult:
+    ) -> AsyncParseResult:
         """Parse conversations from sources via acquire → parse flow.
 
         This is a convenience method that runs both stages:
@@ -134,29 +133,30 @@ class ParsingService:
             progress_callback: Optional callback for progress updates
 
         Returns:
-            ParseResult with counts and processed conversation IDs
+            AsyncParseResult with counts and processed conversation IDs
         """
-        from polylogue.pipeline.services.acquisition import AcquisitionService
+        from polylogue.pipeline.services.async_acquisition import AsyncAcquisitionService
 
         backend = self.repository._backend
         if backend is None:
             raise RuntimeError("Repository backend is not initialized")
 
-        # Stage 1: ACQUIRE - store raw bytes
-        acquire_service = AcquisitionService(backend=backend)
-        acquire_result = acquire_service.acquire_sources(
+        # Stage 1: ACQUIRE - store raw bytes (async)
+        acquire_service = AsyncAcquisitionService(backend=backend)
+        acquire_result = await acquire_service.acquire_sources(
             sources,
             progress_callback=progress_callback,
         )
 
         # Find orphaned raw records (raw data exists but conversation was deleted/missing)
-        conn = backend._get_connection()
-        orphaned_rows = conn.execute("""
-            SELECT r.raw_id
-            FROM raw_conversations r
-            LEFT JOIN conversations c ON r.raw_id = c.raw_id
-            WHERE c.conversation_id IS NULL
-        """).fetchall()
+        async with backend._get_connection() as conn:
+            cursor = await conn.execute("""
+                SELECT r.raw_id
+                FROM raw_conversations r
+                LEFT JOIN conversations c ON r.raw_id = c.raw_id
+                WHERE c.conversation_id IS NULL
+            """)
+            orphaned_rows = await cursor.fetchall()
         orphaned_ids = [row["raw_id"] for row in orphaned_rows]
 
         # Combine newly acquired + orphaned raw IDs
@@ -175,26 +175,26 @@ class ParsingService:
 
         # Stage 2: PARSE - parse raw_conversations into conversations
         if all_raw_ids:
-            return self.parse_from_raw(
+            return await self.parse_from_raw(
                 raw_ids=all_raw_ids,
                 progress_callback=progress_callback,
             )
         else:
             # Nothing to process
-            return ParseResult()
+            return AsyncParseResult()
 
     # Batch size for processing raw records to limit memory usage.
     # Each raw record may contain multi-MB JSONL content; loading thousands
     # at once caused OOM kills on archives with >3000 conversations.
     RAW_BATCH_SIZE = 200
 
-    def parse_from_raw(
+    async def parse_from_raw(
         self,
         *,
         raw_ids: list[str] | None = None,
         provider: str | None = None,
         progress_callback: Any | None = None,
-    ) -> ParseResult:
+    ) -> AsyncParseResult:
         """Parse raw_conversations from DB into conversations.
 
         This is the proper PARSE stage: reads from raw_conversations table
@@ -210,12 +210,12 @@ class ParsingService:
             progress_callback: Optional callback for progress updates
 
         Returns:
-            ParseResult with counts and processed conversation IDs
+            AsyncParseResult with counts and processed conversation IDs
         """
-        result = ParseResult()
+        result = AsyncParseResult()
 
         # Reset per-run drift tracking
-        with self._drift_lock:
+        async with self._drift_lock:
             self._drift_counts.clear()
 
         # Use the repository's backend - same connection management
@@ -228,13 +228,13 @@ class ParsingService:
             ids_to_process = list(raw_ids)
         else:
             ids_to_process = [
-                r.raw_id for r in backend.iter_raw_conversations(provider=provider)
+                r.raw_id async for r in backend.iter_raw_conversations(provider=provider)
             ]
 
         # Process in batches to limit memory
         for batch_start in range(0, len(ids_to_process), self.RAW_BATCH_SIZE):
-            batch_ids = ids_to_process[batch_start:batch_start + self.RAW_BATCH_SIZE]
-            self._process_raw_batch(backend, batch_ids, result, progress_callback)
+            batch_ids = ids_to_process[batch_start : batch_start + self.RAW_BATCH_SIZE]
+            await self._process_raw_batch(backend, batch_ids, result, progress_callback)
 
         if result.processed_ids:
             invalidate_search_cache()
@@ -242,32 +242,28 @@ class ParsingService:
 
         # Copy per-run drift counts to result and check regeneration threshold
         result.drift_counts = dict(self._drift_counts)
-        self._maybe_regenerate_schemas(result.drift_counts)
+        await self._maybe_regenerate_schemas(result.drift_counts)
 
         return result
 
-    def _process_raw_batch(
+    async def _process_raw_batch(
         self,
         backend: Any,
         batch_ids: list[str],
-        result: ParseResult,
+        result: AsyncParseResult,
         progress_callback: Any | None,
     ) -> None:
         """Process a batch of raw conversation IDs."""
         # Load only this batch into memory
-        raw_records = [backend.get_raw_conversation(raw_id) for raw_id in batch_ids]
+        raw_records = [await backend.get_raw_conversation(raw_id) for raw_id in batch_ids]
         raw_records = [r for r in raw_records if r is not None]
-
-        # Ensure main thread's read connection is in a clean state
-        main_conn = backend._get_connection()
-        main_conn.commit()
 
         # Parse raw records into conversation objects
         items_to_process: list[tuple[ParsedConversation, str, str]] = []
 
         for raw_record in raw_records:
             try:
-                parsed_convos = self._parse_raw_record(raw_record)
+                parsed_convos = await self._parse_raw_record(raw_record)
                 source_name = raw_record.source_name or raw_record.source_path
                 for convo in parsed_convos:
                     items_to_process.append((convo, source_name, raw_record.raw_id))
@@ -283,63 +279,41 @@ class ParsingService:
         # Free raw records from memory before processing
         del raw_records
 
-        # Process parsed conversations with thread pool
-        with concurrent.futures.ThreadPoolExecutor(max_workers=4) as executor:
-            futures: dict[concurrent.futures.Future[_ProcessResult], str] = {}
+        # Process parsed conversations with asyncio concurrency control
+        semaphore = asyncio.Semaphore(16)  # Limit concurrent operations
 
-            def _process_one(
-                convo_item: ParsedConversation,
-                source_name_item: str,
-                raw_id: str,
-            ) -> _ProcessResult:
-                thread_conn = backend._get_connection()
-                convo_id, counts, changed = prepare_records(
+        async def _process_one(
+            convo_item: ParsedConversation,
+            source_name_item: str,
+            raw_id: str,
+        ) -> _ProcessResult:
+            async with semaphore:
+                convo_id, counts, changed = await async_prepare_records(
                     convo_item,
                     source_name_item,
                     archive_root=self.archive_root,
-                    conn=thread_conn,
+                    backend=backend,
                     repository=self.repository,
                     raw_id=raw_id,
                 )
                 return (convo_id, counts, changed)
 
-            def _handle_future(fut: concurrent.futures.Future[_ProcessResult]) -> None:
-                convo_id, result_counts, content_changed = fut.result()
-                result.merge_result(convo_id, result_counts, content_changed)
+        # Process all items concurrently with semaphore limiting concurrency
+        tasks = [_process_one(convo, source_name, raw_id) for convo, source_name, raw_id in items_to_process]
+
+        for coro in asyncio.as_completed(tasks):
+            try:
+                convo_id, result_counts, content_changed = await coro
+                await result.merge_result(convo_id, result_counts, content_changed)
+                if progress_callback:
+                    progress_callback(1, desc="Parsing")
+            except Exception as exc:
+                logger.error("Error processing conversation: %s", exc)
+                result.parse_failures += 1
                 if progress_callback:
                     progress_callback(1, desc="Parsing")
 
-            for convo, source_name, raw_id in items_to_process:
-                while len(futures) > 16:
-                    done, _ = concurrent.futures.wait(futures.keys(), return_when=concurrent.futures.FIRST_COMPLETED)
-                    for fut in done:
-                        try:
-                            _handle_future(fut)
-                        except Exception as exc:
-                            conv_id = futures.get(fut, "?")
-                            logger.error("Error processing conversation %s: %s", conv_id, exc)
-                            result.parse_failures += 1
-                            if progress_callback:
-                                progress_callback(1, desc="Parsing")
-                        finally:
-                            del futures[fut]
-
-                future = executor.submit(_process_one, convo, source_name, raw_id)
-                futures[future] = convo.provider_conversation_id
-
-            for fut in concurrent.futures.as_completed(futures):
-                try:
-                    _handle_future(fut)
-                except Exception as exc:
-                    conv_id = futures.get(fut, "?")
-                    logger.error("Error processing conversation %s: %s", conv_id, exc)
-                    result.parse_failures += 1
-                    if progress_callback:
-                        progress_callback(1, desc="Parsing")
-                finally:
-                    del futures[fut]
-
-    def _parse_raw_record(self, raw_record: RawConversationRecord) -> list[ParsedConversation]:
+    async def _parse_raw_record(self, raw_record: RawConversationRecord) -> list[ParsedConversation]:
         """Parse a raw conversation record into ParsedConversation(s).
 
         Handles both single JSON documents and JSONL (newline-delimited JSON).
@@ -373,7 +347,7 @@ class ParsingService:
             payload = lines
 
         # Schema validation: non-blocking, logs drift and errors
-        self._validate_payload(raw_record.provider_name, payload, raw_record.raw_id)
+        await self._validate_payload(raw_record.provider_name, payload, raw_record.raw_id)
 
         # Use the existing parser dispatcher
         return _parse_json_payload(
@@ -382,7 +356,7 @@ class ParsingService:
             raw_record.raw_id,  # Use raw_id as fallback conversation ID
         )
 
-    def _validate_payload(
+    async def _validate_payload(
         self,
         provider_name: str,
         payload: Any,
@@ -428,10 +402,8 @@ class ParsingService:
                     raw_id=raw_id,
                     drift=result.drift_warnings[:10],
                 )
-                with self._drift_lock:
-                    self._drift_counts[provider_name] = (
-                        self._drift_counts.get(provider_name, 0) + 1
-                    )
+                async with self._drift_lock:
+                    self._drift_counts[provider_name] = self._drift_counts.get(provider_name, 0) + 1
         except Exception as exc:
             logger.debug(
                 "Schema validation skipped for %s: %s",
@@ -443,7 +415,7 @@ class ParsingService:
     # many drift warnings in a single parsing run.
     DRIFT_REGEN_THRESHOLD = 5
 
-    def _maybe_regenerate_schemas(self, drift_counts: dict[str, int]) -> None:
+    async def _maybe_regenerate_schemas(self, drift_counts: dict[str, int]) -> None:
         """Auto-regenerate schemas for providers with excessive drift.
 
         If a provider accumulated more than DRIFT_REGEN_THRESHOLD drift
@@ -453,10 +425,7 @@ class ParsingService:
         from polylogue.schemas.registry import SchemaRegistry
         from polylogue.schemas.schema_inference import generate_provider_schema
 
-        providers_to_regen = [
-            p for p, count in drift_counts.items()
-            if count > self.DRIFT_REGEN_THRESHOLD
-        ]
+        providers_to_regen = [p for p, count in drift_counts.items() if count > self.DRIFT_REGEN_THRESHOLD]
 
         if not providers_to_regen:
             return
@@ -489,3 +458,6 @@ class ParsingService:
                     provider,
                     exc,
                 )
+
+
+__all__ = ["AsyncParsingService", "AsyncParseResult"]
