+++ polylogue/storage/repository.py
@@ -1,105 +1,193 @@
-"""Unified repository for conversation read/write operations."""
+"""Async storage repository for conversation persistence.
+
+Provides async/await interface for storing and retrieving conversations.
+Wraps AsyncSQLiteBackend for parallel operations.
+
+All methods are async and use eager loading (Conversation.from_records)
+instead of lazy loading, since async I/O already enables efficient parallel
+fetching of conversations, messages, and attachments together.
+"""
 
 from __future__ import annotations
 
+import asyncio
 import builtins
-from collections.abc import Iterator
+from collections.abc import AsyncIterator
 from typing import TYPE_CHECKING
 
 from polylogue.lib.log import get_logger
-from polylogue.lib.messages import MessageSource
 from polylogue.lib.models import Conversation, ConversationSummary, Message
-from polylogue.storage.backends.sqlite import SQLiteBackend
+from polylogue.storage.backends.async_sqlite import AsyncSQLiteBackend
+from polylogue.storage.store import AttachmentRecord, ConversationRecord, MessageRecord, RunRecord
 from polylogue.types import ConversationId
 
-from .store import (
-    AttachmentRecord,
-    ConversationRecord,
-    MessageRecord,
-    RunRecord,
-)
-
 if TYPE_CHECKING:
+    from pathlib import Path
+
+    from polylogue.lib import filters
     from polylogue.lib.stats import ArchiveStats
     from polylogue.protocols import VectorProvider
 
 logger = get_logger(__name__)
 
 
-class RepositoryMessageSource(MessageSource):
-    """Adapter providing MessageSource protocol for lazy message loading."""
+class AsyncConversationRepository:
+    """Async repository for conversation storage operations.
 
-    def __init__(self, backend: SQLiteBackend) -> None:
-        self._backend = backend
+    Wraps AsyncSQLiteBackend to provide high-level async storage interface with
+    full feature parity to sync ConversationRepository.
 
-    def iter_messages(self, conversation_id: str) -> Iterator[Message]:
-        for record in self._backend.iter_messages(conversation_id):
-            yield Message.from_record(record, attachments=[])
-
-    def count_messages(self, conversation_id: str) -> int:
-        stats = self._backend.get_conversation_stats(conversation_id)
-        return stats.get("total_messages", 0)
-
-
-class ConversationRepository:
-    """Unified repository for conversation read/write operations.
+    All methods are async. Eager loading (Conversation.from_records) is used
+    for fetching conversations, enabling efficient parallel I/O via asyncio.gather()
+    for conversations, messages, and attachments.
 
     Write safety is provided by SQLite's ``BEGIN IMMEDIATE`` transactions
-    in the backend layer (combined with ``busy_timeout=30000``), not by a
-    Python-level lock.  This allows concurrent reads from other threads
-    to proceed unblocked during writes.
+    in the backend layer, combined with asyncio.Lock() serialization.
+
+    Example:
+        async with AsyncConversationRepository() as repo:
+            conv = await repo.get("claude:abc123")
+            convs = await repo.list(limit=10)
+            await repo.save_conversation(conv_rec, msgs, atts)
     """
 
-    def __init__(self, backend: SQLiteBackend) -> None:
-        self._backend = backend
-        # Store db_path for thread workers in ParsingService
-        self._db_path = getattr(backend, "_db_path", None)
+    def __init__(
+        self,
+        backend: AsyncSQLiteBackend | None = None,
+        db_path: Path | None = None,
+    ) -> None:
+        """Initialize async storage repository.
+
+        Args:
+            backend: Optional AsyncSQLiteBackend instance. If provided, db_path is ignored.
+            db_path: Optional path to database file. Used if backend is None.
+        """
+        if backend is not None:
+            self._backend = backend
+        else:
+            self._backend = AsyncSQLiteBackend(db_path=db_path)
+
+        # Store db_path for compatibility with sync repository (used by message mapping)
+        self._db_path = getattr(self._backend, "_db_path", None)
+
+    async def __aenter__(self) -> AsyncConversationRepository:
+        """Enter async context manager."""
+        return self
+
+    async def __aexit__(self, exc_type: object, exc_val: object, exc_tb: object) -> None:
+        """Exit async context manager."""
+        await self.close()
 
     @property
-    def backend(self) -> SQLiteBackend:
-        """Access the underlying storage backend."""
+    def backend(self) -> AsyncSQLiteBackend:
+        """Access the underlying async storage backend."""
         return self._backend
+
+    async def close(self) -> None:
+        """Close database connections and release resources."""
+        await self._backend.close()
 
     # --- Read Methods ---
 
-    def resolve_id(self, id_prefix: str) -> ConversationId | None:
-        """Resolve a partial ID to a full conversation ID."""
-        resolved = self._backend.resolve_id(id_prefix)
+    async def resolve_id(self, id_prefix: str) -> ConversationId | None:
+        """Resolve a partial ID to a full conversation ID.
+
+        Args:
+            id_prefix: Full or partial conversation ID
+
+        Returns:
+            Full ConversationId if resolved, None if ambiguous or not found
+        """
+        resolved = await self._backend.resolve_id(id_prefix)
         return ConversationId(resolved) if resolved else None
 
-    def get(self, conversation_id: str) -> Conversation | None:
-        """Get a conversation with lazy message loading."""
-        conv_record = self._backend.get_conversation(conversation_id)
+    async def get(self, conversation_id: str) -> Conversation | None:
+        """Get a conversation with all messages and attachments (eager loading).
+
+        Fetches conversation, messages, and attachments in parallel.
+
+        Args:
+            conversation_id: Full conversation ID
+
+        Returns:
+            Conversation with all data loaded, or None if not found
+        """
+        conv_record = await self._backend.get_conversation(conversation_id)
         if not conv_record:
             return None
 
-        source = RepositoryMessageSource(self._backend)
-        return Conversation.from_lazy(conv_record, source)
-
-    def view(self, conversation_id: str) -> Conversation | None:
-        """Get a conversation with ID resolution support."""
-        full_id = self.resolve_id(conversation_id) or conversation_id
-        return self.get(full_id)
-
-    def get_eager(self, conversation_id: str) -> Conversation | None:
-        """Get a conversation with all messages loaded."""
-        conv_record = self._backend.get_conversation(conversation_id)
-        if not conv_record:
-            return None
-
-        msg_records = self._backend.get_messages(conversation_id)
-        att_records = self._backend.get_attachments(conversation_id)
+        # Fetch messages and attachments in parallel
+        msg_records, att_records = await asyncio.gather(
+            self._backend.get_messages(conversation_id),
+            self._backend.get_attachments(conversation_id),
+        )
 
         return Conversation.from_records(conv_record, msg_records, att_records)
 
-    def get_summary(self, conversation_id: str) -> ConversationSummary | None:
-        """Get a single conversation summary without loading messages."""
-        conv_record = self._backend.get_conversation(conversation_id)
+    async def view(self, conversation_id: str) -> Conversation | None:
+        """Get a conversation with ID resolution support.
+
+        Attempts to resolve a partial ID to a full ID, then fetches the conversation.
+
+        Args:
+            conversation_id: Full or partial conversation ID
+
+        Returns:
+            Conversation with all data loaded, or None if not found
+        """
+        full_id = await self.resolve_id(conversation_id) or conversation_id
+        return await self.get(str(full_id))
+
+    async def get_eager(self, conversation_id: str) -> Conversation | None:
+        """Get a conversation with all messages and attachments.
+
+        Alias for get() - all async operations are eager by design.
+
+        Args:
+            conversation_id: Full conversation ID
+
+        Returns:
+            Conversation with all data loaded, or None if not found
+        """
+        return await self.get(conversation_id)
+
+    async def get_conversation(self, conversation_id: str) -> ConversationRecord | None:
+        """Get conversation record by ID (backend method exposed for compatibility).
+
+        Args:
+            conversation_id: Full conversation ID
+
+        Returns:
+            ConversationRecord or None if not found
+        """
+        return await self._backend.get_conversation(conversation_id)
+
+    async def conversation_exists(self, content_hash: str) -> bool:
+        """Check if conversation with given content hash exists.
+
+        Args:
+            content_hash: SHA-256 hash of conversation content
+
+        Returns:
+            True if conversation exists, False otherwise
+        """
+        return await self._backend.conversation_exists_by_hash(content_hash)
+
+    async def get_summary(self, conversation_id: str) -> ConversationSummary | None:
+        """Get a conversation summary without loading messages.
+
+        Args:
+            conversation_id: Full conversation ID
+
+        Returns:
+            ConversationSummary, or None if not found
+        """
+        conv_record = await self._backend.get_conversation(conversation_id)
         if not conv_record:
             return None
         return ConversationSummary.from_record(conv_record)
 
-    def list_summaries(
+    async def list_summaries(
         self,
         limit: int | None = 50,
         offset: int = 0,
@@ -110,8 +198,22 @@ class ConversationRepository:
         until: str | None = None,
         title_contains: str | None = None,
     ) -> builtins.list[ConversationSummary]:
-        """List conversation summaries without loading messages."""
-        conv_records = self._backend.list_conversations(
+        """List conversation summaries without loading messages.
+
+        Args:
+            limit: Maximum number of results
+            offset: Number of results to skip
+            provider: Filter by single provider (backwards compat)
+            providers: Filter by multiple providers
+            source: Filter by source name
+            since: Filter to conversations updated on/after this ISO date
+            until: Filter to conversations updated on/before this ISO date
+            title_contains: Filter by title substring (case-insensitive)
+
+        Returns:
+            List of ConversationSummary objects
+        """
+        conv_records = await self._backend.list_conversations(
             source=source,
             provider=provider,
             providers=providers,
@@ -123,7 +225,7 @@ class ConversationRepository:
         )
         return [ConversationSummary.from_record(rec) for rec in conv_records]
 
-    def list(
+    async def list(
         self,
         limit: int | None = 50,
         offset: int = 0,
@@ -133,8 +235,21 @@ class ConversationRepository:
         until: str | None = None,
         title_contains: str | None = None,
     ) -> builtins.list[Conversation]:
-        """List conversations with lazy message loading."""
-        conv_records = self._backend.list_conversations(
+        """List conversations with eager-loaded messages and attachments.
+
+        Args:
+            limit: Maximum number of results
+            offset: Number of results to skip
+            provider: Filter by single provider (backwards compat)
+            providers: Filter by multiple providers
+            since: Filter to conversations updated on/after this ISO date
+            until: Filter to conversations updated on/before this ISO date
+            title_contains: Filter by title substring (case-insensitive)
+
+        Returns:
+            List of Conversation objects with all data eager-loaded
+        """
+        conv_records = await self._backend.list_conversations(
             provider=provider,
             providers=providers,
             limit=limit,
@@ -143,10 +258,14 @@ class ConversationRepository:
             until=until,
             title_contains=title_contains,
         )
-        source = RepositoryMessageSource(self._backend)
-        return [Conversation.from_lazy(rec, source) for rec in conv_records]
 
-    def count(
+        # Fetch messages and attachments for all conversations in parallel
+        if not conv_records:
+            return []
+
+        return await self._get_many([rec.conversation_id for rec in conv_records])
+
+    async def count(
         self,
         provider: str | None = None,
         providers: builtins.list[str] | None = None,
@@ -154,8 +273,19 @@ class ConversationRepository:
         until: str | None = None,
         title_contains: str | None = None,
     ) -> int:
-        """Count conversations matching filters without loading data."""
-        return self._backend.count_conversations(
+        """Count conversations matching filters.
+
+        Args:
+            provider: Filter by single provider
+            providers: Filter by multiple providers
+            since: Filter to conversations updated on/after this ISO date
+            until: Filter to conversations updated on/before this ISO date
+            title_contains: Filter by title substring
+
+        Returns:
+            Count of matching conversations
+        """
+        return await self._backend.count_conversations(
             provider=provider,
             providers=providers,
             since=since,
@@ -163,35 +293,79 @@ class ConversationRepository:
             title_contains=title_contains,
         )
 
-    def get_parent(self, conversation_id: str) -> Conversation | None:
-        """Get the parent conversation if it exists."""
-        conv = self.get(conversation_id)
+    async def get_source_conversations(self, provider: str) -> builtins.list[str]:
+        """Get all conversation IDs for a given provider.
+
+        Args:
+            provider: Provider name to filter by
+
+        Returns:
+            List of conversation IDs from that provider
+        """
+        records = await self._backend.list_conversations(provider=provider, limit=None)
+        return [rec.conversation_id for rec in records]
+
+    async def get_parent(self, conversation_id: str) -> Conversation | None:
+        """Get the parent conversation if it exists.
+
+        Args:
+            conversation_id: Child conversation ID
+
+        Returns:
+            Parent Conversation, or None if no parent
+        """
+        conv = await self.get(conversation_id)
         if conv and conv.parent_id:
-            return self.get(str(conv.parent_id))
+            return await self.get(str(conv.parent_id))
         return None
 
-    def get_children(self, conversation_id: str) -> builtins.list[Conversation]:
-        """Get all direct children of this conversation."""
-        child_records = self._backend.list_conversations(parent_id=conversation_id)
-        source = RepositoryMessageSource(self._backend)
-        return [Conversation.from_lazy(rec, source) for rec in child_records]
+    async def get_children(self, conversation_id: str) -> builtins.list[Conversation]:
+        """Get all direct children of this conversation.
 
-    def get_root(self, conversation_id: str) -> Conversation:
-        """Walk up to find the root conversation."""
-        current = self.get(conversation_id)
+        Args:
+            conversation_id: Parent conversation ID
+
+        Returns:
+            List of child Conversation objects with all data eager-loaded
+        """
+        child_records = await self._backend.list_conversations(parent_id=conversation_id)
+        if not child_records:
+            return []
+        return await self._get_many([rec.conversation_id for rec in child_records])
+
+    async def get_root(self, conversation_id: str) -> Conversation:
+        """Walk up the parent chain to find the root conversation.
+
+        Args:
+            conversation_id: Any conversation ID in the tree
+
+        Returns:
+            The root Conversation object
+
+        Raises:
+            ValueError: If the conversation is not found
+        """
+        current = await self.get(conversation_id)
         if not current:
             raise ValueError(f"Conversation {conversation_id} not found")
 
         while current.parent_id:
-            parent = self.get(str(current.parent_id))
+            parent = await self.get(str(current.parent_id))
             if not parent:
                 break
             current = parent
         return current
 
-    def get_session_tree(self, conversation_id: str) -> builtins.list[Conversation]:
-        """Get all conversations in the session tree."""
-        root = self.get_root(conversation_id)
+    async def get_session_tree(self, conversation_id: str) -> builtins.list[Conversation]:
+        """Get all conversations in the session tree (root + all descendants).
+
+        Args:
+            conversation_id: Any conversation in the tree
+
+        Returns:
+            List of all Conversation objects in tree (breadth-first order)
+        """
+        root = await self.get_root(conversation_id)
 
         tree = []
         queue = [root]
@@ -199,67 +373,155 @@ class ConversationRepository:
         while queue:
             current = queue.pop(0)
             tree.append(current)
-            children = self.get_children(str(current.id))
+            children = await self.get_children(str(current.id))
             queue.extend(children)
 
         return tree
 
-    def search_summaries(
-        self, query: str, limit: int = 20, providers: builtins.list[str] | None = None
+    async def search_summaries(
+        self,
+        query: str,
+        limit: int = 20,
+        providers: builtins.list[str] | None = None,
     ) -> builtins.list[ConversationSummary]:
-        """Search conversations and return summaries."""
-        ids = self._backend.search_conversations(query, limit=limit, providers=providers)
-        summaries = []
-        for cid in ids:
-            record = self._backend.get_conversation(cid)
-            if record:
-                summaries.append(ConversationSummary.from_record(record))
-        return summaries
+        """Search conversations and return summaries.
 
-    def search(
-        self, query: str, limit: int = 20, providers: builtins.list[str] | None = None
+        Args:
+            query: Full-text search query
+            limit: Maximum results
+            providers: Optional list of provider names to filter by
+
+        Returns:
+            List of ConversationSummary objects matching the query
+        """
+        ids = await self._backend.search_conversations(query, limit=limit, providers=providers)
+        if not ids:
+            return []
+
+        # Fetch all matching conversations in parallel
+        records = await self._backend.get_conversations_batch(ids)
+        return [ConversationSummary.from_record(rec) for rec in records]
+
+    async def search(
+        self,
+        query: str,
+        limit: int = 20,
+        providers: builtins.list[str] | None = None,
     ) -> builtins.list[Conversation]:
-        """Search conversations using full-text search."""
-        ids = self._backend.search_conversations(query, limit=limit, providers=providers)
-        return self._get_many(ids)
+        """Search conversations using full-text search.
 
-    def _get_many(self, conversation_ids: builtins.list[str]) -> builtins.list[Conversation]:
-        """Bulk fetch lazy conversation objects in a single SQL query."""
+        Args:
+            query: Full-text search query
+            limit: Maximum results
+            providers: Optional list of provider names to filter by
+
+        Returns:
+            List of Conversation objects with all data eager-loaded
+        """
+        ids = await self._backend.search_conversations(query, limit=limit, providers=providers)
+        return await self._get_many(ids)
+
+    async def _get_many(self, conversation_ids: builtins.list[str]) -> builtins.list[Conversation]:
+        """Bulk fetch conversations with eager-loaded messages and attachments.
+
+        Uses asyncio.gather() to fetch messages and attachments in parallel
+        for all conversations.
+
+        Args:
+            conversation_ids: List of conversation IDs to fetch
+
+        Returns:
+            List of Conversation objects in the order of input IDs
+        """
         if not conversation_ids:
             return []
-        records = self._backend.get_conversations_batch(conversation_ids)
-        source = RepositoryMessageSource(self._backend)
-        return [Conversation.from_lazy(rec, source) for rec in records]
 
-    def get_conversation_stats(self, conversation_id: str) -> dict[str, int] | None:
-        """Get message counts without loading messages."""
-        conv_record = self._backend.get_conversation(conversation_id)
+        # Fetch conversation records
+        records = await self._backend.get_conversations_batch(conversation_ids)
+        if not records:
+            return []
+
+        # Build dict for order preservation
+        by_id = {rec.conversation_id: rec for rec in records}
+
+        # Fetch all messages and attachments in parallel
+        fetch_tasks = [
+            (cid, asyncio.gather(
+                self._backend.get_messages(cid),
+                self._backend.get_attachments(cid),
+            ))
+            for cid in conversation_ids
+            if cid in by_id
+        ]
+
+        results: builtins.list[Conversation] = []
+        for cid, gather_task in fetch_tasks:
+            msg_records, att_records = await gather_task
+            conv_record = by_id[cid]
+            results.append(Conversation.from_records(conv_record, msg_records, att_records))
+
+        return results
+
+    async def get_conversation_stats(self, conversation_id: str) -> dict[str, int] | None:
+        """Get message counts without loading messages.
+
+        Args:
+            conversation_id: Conversation ID
+
+        Returns:
+            Dict with counts (total_messages, dialogue_messages, tool_messages),
+            or None if conversation not found
+        """
+        conv_record = await self._backend.get_conversation(conversation_id)
         if not conv_record:
             return None
-        return self._backend.get_conversation_stats(conversation_id)
+        return await self._backend.get_conversation_stats(conversation_id)
 
-    def iter_messages(
+    async def iter_messages(
         self,
         conversation_id: str,
         *,
         dialogue_only: bool = False,
         limit: int | None = None,
-    ) -> Iterator[Message]:
-        """Stream messages without loading full conversation."""
-        for record in self._backend.iter_messages(
+    ) -> AsyncIterator[Message]:
+        """Stream messages without loading full conversation.
+
+        Memory-efficient iteration for large conversations.
+
+        Args:
+            conversation_id: Conversation ID
+            dialogue_only: If True, only yield user/assistant messages
+            limit: Maximum messages to yield
+
+        Yields:
+            Message objects one at a time
+        """
+        async for record in self._backend.iter_messages(
             conversation_id,
             dialogue_only=dialogue_only,
             limit=limit,
         ):
             yield Message.from_record(record, attachments=[])
 
-    def search_similar(
+    async def search_similar(
         self,
         text: str,
         limit: int = 10,
         vector_provider: VectorProvider | None = None,
     ) -> builtins.list[Conversation]:
-        """Search by semantic similarity."""
+        """Search by semantic similarity.
+
+        Args:
+            text: Query text
+            limit: Maximum results
+            vector_provider: Vector provider for embedding
+
+        Returns:
+            List of similar Conversation objects
+
+        Raises:
+            ValueError: If no vector provider available
+        """
         if not vector_provider:
             raise ValueError("Semantic search requires a vector provider.")
 
@@ -268,7 +530,7 @@ class ConversationRepository:
             return []
 
         message_ids = [msg_id for msg_id, _ in results]
-        msg_to_conv = self._get_message_conversation_mapping(message_ids)
+        msg_to_conv = await self._get_message_conversation_mapping(message_ids)
 
         conv_scores: dict[str, float] = {}
         for msg_id, distance in results:
@@ -282,35 +544,129 @@ class ConversationRepository:
             key=lambda x: conv_scores[x],
         )[:limit]
 
-        return self._get_many(ranked_ids)
+        return await self._get_many(ranked_ids)
 
-    def _get_message_conversation_mapping(self, message_ids: builtins.list[str]) -> dict[str, str]:
-        from polylogue.storage.backends.sqlite import open_connection
+    async def _get_message_conversation_mapping(
+        self, message_ids: builtins.list[str]
+    ) -> dict[str, str]:
+        """Fetch mapping of message IDs to conversation IDs.
 
-        with open_connection(self._db_path) as conn:
-            placeholders = ",".join("?" * len(message_ids))
-            query = f"SELECT message_id, conversation_id FROM messages WHERE message_id IN ({placeholders})"
-            rows = conn.execute(query, message_ids).fetchall()
+        Args:
+            message_ids: List of message IDs
+
+        Returns:
+            Dict mapping message_id -> conversation_id
+        """
+        if not message_ids:
+            return {}
+
+        placeholders = ",".join("?" * len(message_ids))
+        query = f"SELECT message_id, conversation_id FROM messages WHERE message_id IN ({placeholders})"
+
+        async with self._backend._get_connection() as conn:
+            cursor = await conn.execute(query, message_ids)
+            rows = await cursor.fetchall()
+
         return {row["message_id"]: row["conversation_id"] for row in rows}
+
+    def filter(self) -> filters.ConversationFilter:
+        """Create a filter builder for chainable queries.
+
+        Terminal methods (list, first, count, etc.) are async and must be awaited.
+
+        Returns:
+            ConversationFilter instance
+        """
+        from polylogue.lib import filters
+
+        return filters.ConversationFilter(self)
 
     # --- Write Methods ---
 
-    def save_conversation(
+    async def save_conversation(
         self,
-        *,
-        conversation: ConversationRecord,
+        conversation: Conversation | ConversationRecord,
         messages: builtins.list[MessageRecord],
         attachments: builtins.list[AttachmentRecord],
     ) -> dict[str, int]:
-        """Save a conversation with its messages and attachments atomically."""
-        return self._save_via_backend(conversation, messages, attachments)
+        """Save a conversation with its messages and attachments atomically.
 
-    def _save_via_backend(
+        Args:
+            conversation: Conversation model or ConversationRecord to save
+            messages: List of message records
+            attachments: List of attachment records
+
+        Returns:
+            Dictionary with counts:
+                - conversations: New conversations
+                - messages: New messages
+                - attachments: New attachments
+                - skipped_conversations: Unchanged conversations
+                - skipped_messages: Unchanged messages
+                - skipped_attachments: Existing attachments
+        """
+        # Convert Conversation model to ConversationRecord if needed
+        if isinstance(conversation, Conversation):
+            conv_record = self._conversation_to_record(conversation)
+        else:
+            conv_record = conversation
+
+        return await self._save_via_backend(conv_record, messages, attachments)
+
+    def _conversation_to_record(self, conversation: Conversation) -> ConversationRecord:
+        """Convert a Conversation model to a ConversationRecord.
+
+        Args:
+            conversation: Conversation model
+
+        Returns:
+            ConversationRecord
+        """
+        from typing import cast
+
+        from polylogue.types import ContentHash, ConversationId
+
+        created_at_str = conversation.created_at.isoformat() if conversation.created_at else None
+        updated_at_str = conversation.updated_at.isoformat() if conversation.updated_at else (created_at_str or None)
+
+        # Try to extract provider_id from canonical id (format: provider:id)
+        # Fallback to whole ID if pattern doesn't match
+        provider_id = str(conversation.id)
+        if ":" in provider_id and conversation.provider:
+            prefix = f"{conversation.provider}:"
+            if provider_id.startswith(prefix):
+                provider_id = provider_id[len(prefix) :]
+
+        return ConversationRecord(
+            conversation_id=cast(ConversationId, str(conversation.id)),
+            provider_name=conversation.provider,
+            provider_conversation_id=provider_id,
+            title=conversation.title or "",
+            created_at=created_at_str,
+            updated_at=updated_at_str,
+            content_hash=cast(ContentHash, conversation.metadata.get("content_hash", "")),
+            provider_meta=cast(dict[str, object], conversation.metadata.get("provider_meta", {})),
+            metadata=conversation.metadata,
+        )
+
+    async def _save_via_backend(
         self,
         conversation: ConversationRecord,
         messages: builtins.list[MessageRecord],
         attachments: builtins.list[AttachmentRecord],
     ) -> dict[str, int]:
+        """Internal implementation of save_conversation via backend.
+
+        Handles transaction control and skip detection.
+
+        Args:
+            conversation: Conversation record
+            messages: List of message records
+            attachments: List of attachment records
+
+        Returns:
+            Dictionary with operation counts
+        """
         counts: dict[str, int] = {
             "conversations": 0,
             "messages": 0,
@@ -324,16 +680,19 @@ class ConversationRepository:
         if backend is None:
             raise RuntimeError("Backend is not initialized")
 
-        backend.begin()
-        try:
-            existing = backend.get_conversation(conversation.conversation_id)
-            backend.save_conversation(conversation)
+        async with backend.transaction():
+            # Check existing conversation
+            existing = await backend.get_conversation(conversation.conversation_id)
+            await backend.save_conversation_record(conversation)
             if existing and existing.content_hash == conversation.content_hash:
                 counts["skipped_conversations"] += 1
             else:
                 counts["conversations"] += 1
 
-            existing_messages = {msg.message_id: msg for msg in backend.get_messages(conversation.conversation_id)}
+            # Check existing messages
+            existing_messages = {
+                msg.message_id: msg for msg in await backend.get_messages(conversation.conversation_id)
+            }
             for message in messages:
                 existing_msg = existing_messages.get(message.message_id)
                 if existing_msg and existing_msg.content_hash == message.content_hash:
@@ -342,10 +701,12 @@ class ConversationRepository:
                     counts["messages"] += 1
 
             if messages:
-                backend.save_messages(messages)
+                await backend.save_messages(messages)
 
+            # Check existing attachments
             existing_attachments = {
-                att.attachment_id: att for att in backend.get_attachments(conversation.conversation_id)
+                att.attachment_id: att
+                for att in await backend.get_attachments(conversation.conversation_id)
             }
             for attachment in attachments:
                 if attachment.attachment_id in existing_attachments:
@@ -354,52 +715,107 @@ class ConversationRepository:
                     counts["attachments"] += 1
 
             new_attachment_ids: set[str] = {str(att.attachment_id) for att in attachments}
-            backend.prune_attachments(conversation.conversation_id, new_attachment_ids)
+            await backend.prune_attachments(conversation.conversation_id, new_attachment_ids)
 
             if attachments:
-                backend.save_attachments(attachments)
-
-            backend.commit()
-        except Exception:
-            backend.rollback()
-            raise
+                await backend.save_attachments(attachments)
 
         return counts
 
-    def record_run(self, record: RunRecord) -> None:
-        """Record a pipeline run audit entry."""
-        self._backend.record_run(record)
+    async def record_run(self, record: RunRecord) -> None:
+        """Record a pipeline run audit entry.
+
+        Args:
+            record: Run record to persist
+        """
+        await self._backend.record_run(record)
 
     # --- Metadata CRUD ---
 
-    def get_metadata(self, conversation_id: str) -> dict[str, object]:
-        return self._backend.get_metadata(conversation_id)
+    async def get_metadata(self, conversation_id: str) -> dict[str, object]:
+        """Get metadata dictionary for a conversation.
 
-    def update_metadata(self, conversation_id: str, key: str, value: object) -> None:
-        self._backend.update_metadata(conversation_id, key, value)
+        Args:
+            conversation_id: Conversation ID
 
-    def delete_metadata(self, conversation_id: str, key: str) -> None:
-        self._backend.delete_metadata(conversation_id, key)
+        Returns:
+            Metadata dictionary (empty if not found)
+        """
+        return await self._backend.get_metadata(conversation_id)
 
-    def add_tag(self, conversation_id: str, tag: str) -> None:
-        self._backend.add_tag(conversation_id, tag)
+    async def update_metadata(self, conversation_id: str, key: str, value: object) -> None:
+        """Set a single metadata key.
 
-    def remove_tag(self, conversation_id: str, tag: str) -> None:
-        self._backend.remove_tag(conversation_id, tag)
+        Args:
+            conversation_id: Conversation ID
+            key: Metadata key
+            value: Value to set
+        """
+        await self._backend.update_metadata(conversation_id, key, value)
 
-    def list_tags(self, *, provider: str | None = None) -> dict[str, int]:
-        """List all tags with counts."""
-        return self._backend.list_tags(provider=provider)
+    async def delete_metadata(self, conversation_id: str, key: str) -> None:
+        """Remove a metadata key.
 
-    def set_metadata(self, conversation_id: str, metadata: dict[str, object]) -> None:
-        self._backend.set_metadata(conversation_id, metadata)
+        Args:
+            conversation_id: Conversation ID
+            key: Metadata key to delete
+        """
+        await self._backend.delete_metadata(conversation_id, key)
 
-    def delete_conversation(self, conversation_id: str) -> bool:
-        return self._backend.delete_conversation(conversation_id)
+    async def add_tag(self, conversation_id: str, tag: str) -> None:
+        """Add a tag to the conversation's tags list.
+
+        Args:
+            conversation_id: Conversation ID
+            tag: Tag to add
+        """
+        await self._backend.add_tag(conversation_id, tag)
+
+    async def remove_tag(self, conversation_id: str, tag: str) -> None:
+        """Remove a tag from the conversation's tags list.
+
+        Args:
+            conversation_id: Conversation ID
+            tag: Tag to remove
+        """
+        await self._backend.remove_tag(conversation_id, tag)
+
+    async def list_tags(self, *, provider: str | None = None) -> dict[str, int]:
+        """List all tags with counts.
+
+        Args:
+            provider: Optional provider filter
+
+        Returns:
+            Dict of tag -> count, sorted by count descending
+        """
+        return await self._backend.list_tags(provider=provider)
+
+    async def set_metadata(self, conversation_id: str, metadata: dict[str, object]) -> None:
+        """Replace entire metadata dictionary.
+
+        Args:
+            conversation_id: Conversation ID
+            metadata: New metadata dictionary
+        """
+        await self._backend.set_metadata(conversation_id, metadata)
+
+    async def delete_conversation(self, conversation_id: str) -> bool:
+        """Delete conversation and all related records.
+
+        Removes conversation, messages, attachment refs, and FTS index entries.
+
+        Args:
+            conversation_id: Conversation ID to delete
+
+        Returns:
+            True if deleted, False if not found
+        """
+        return await self._backend.delete_conversation(conversation_id)
 
     # --- Vector Search Methods ---
 
-    def embed_conversation(
+    async def embed_conversation(
         self,
         conversation_id: str,
         vector_provider: VectorProvider | None = None,
@@ -418,19 +834,20 @@ class ConversationRepository:
         """
         if vector_provider is None:
             from polylogue.storage.search_providers import create_vector_provider
+
             vector_provider = create_vector_provider()
 
         if vector_provider is None:
             raise ValueError("No vector provider available. Set VOYAGE_API_KEY.")
 
-        messages = self._backend.get_messages(conversation_id)
+        messages = await self._backend.get_messages(conversation_id)
         if not messages:
             return 0
 
         vector_provider.upsert(conversation_id, messages)
         return len(messages)
 
-    def similarity_search(
+    async def similarity_search(
         self,
         query: str,
         limit: int = 10,
@@ -445,9 +862,13 @@ class ConversationRepository:
 
         Returns:
             List of (conversation_id, message_id, distance) tuples
+
+        Raises:
+            ValueError: If no vector provider configured
         """
         if vector_provider is None:
             from polylogue.storage.search_providers import create_vector_provider
+
             vector_provider = create_vector_provider()
 
         if vector_provider is None:
@@ -457,9 +878,9 @@ class ConversationRepository:
         if not results:
             return []
 
-        # Batch lookup conversation IDs (single query instead of N+1)
+        # Batch lookup conversation IDs
         message_ids = [msg_id for msg_id, _ in results]
-        msg_to_conv = self._get_message_conversation_mapping(message_ids)
+        msg_to_conv = await self._get_message_conversation_mapping(message_ids)
 
         return [
             (msg_to_conv[msg_id], msg_id, distance)
@@ -467,7 +888,7 @@ class ConversationRepository:
             if msg_id in msg_to_conv
         ]
 
-    def get_archive_stats(self) -> ArchiveStats:
+    async def get_archive_stats(self) -> ArchiveStats:
         """Get comprehensive archive statistics.
 
         Returns:
@@ -475,46 +896,47 @@ class ConversationRepository:
         """
         from polylogue.lib.stats import ArchiveStats
 
-        conn = self._backend._get_connection()
+        async with self._backend._get_connection() as conn:
+            # Total counts
+            cursor = await conn.execute("SELECT COUNT(*) FROM conversations")
+            conv_count = (await cursor.fetchone())[0]
 
-        conv_count = conn.execute(
-            "SELECT COUNT(*) FROM conversations"
-        ).fetchone()[0]
+            cursor = await conn.execute("SELECT COUNT(*) FROM messages")
+            msg_count = (await cursor.fetchone())[0]
 
-        msg_count = conn.execute(
-            "SELECT COUNT(*) FROM messages"
-        ).fetchone()[0]
+            cursor = await conn.execute("SELECT COUNT(*) FROM attachments")
+            att_count = (await cursor.fetchone())[0]
 
-        att_count = conn.execute(
-            "SELECT COUNT(*) FROM attachments"
-        ).fetchone()[0]
+            # Provider breakdown
+            cursor = await conn.execute(
+                """
+                SELECT provider_name, COUNT(*) as count
+                FROM conversations
+                GROUP BY provider_name
+                """
+            )
+            provider_rows = await cursor.fetchall()
+            providers = {row["provider_name"]: row["count"] for row in provider_rows}
 
-        provider_rows = conn.execute(
-            """
-            SELECT provider_name, COUNT(*) as count
-            FROM conversations
-            GROUP BY provider_name
-            """
-        ).fetchall()
-        providers = {row["provider_name"]: row["count"] for row in provider_rows}
+            # Check embedding status if table exists
+            embedded_convs = 0
+            embedded_msgs = 0
+            try:
+                cursor = await conn.execute(
+                    "SELECT COUNT(*) FROM embedding_status WHERE needs_reindex = 0"
+                )
+                embedded_convs = (await cursor.fetchone())[0]
 
-        # Check embedding status if table exists
-        embedded_convs = 0
-        embedded_msgs = 0
-        try:
-            embedded_convs = conn.execute(
-                "SELECT COUNT(*) FROM embedding_status WHERE needs_reindex = 0"
-            ).fetchone()[0]
-            embedded_msgs = conn.execute(
-                "SELECT COUNT(*) FROM message_embeddings"
-            ).fetchone()[0]
-        except Exception as exc:
-            logger.warning("Embedding stats query failed: %s", exc)
+                cursor = await conn.execute("SELECT COUNT(*) FROM message_embeddings")
+                embedded_msgs = (await cursor.fetchone())[0]
+            except Exception as exc:
+                logger.warning("Embedding stats query failed: %s", exc)
 
         # Get database size
         db_size = 0
         try:
             from pathlib import Path
+
             db_size = Path(self._db_path).stat().st_size if self._db_path else 0
         except Exception as exc:
             logger.warning("DB size check failed: %s", exc)
@@ -530,16 +952,4 @@ class ConversationRepository:
         )
 
 
-def _records_to_conversation(
-    conversation: ConversationRecord,
-    messages: list[MessageRecord],
-    attachments: list[AttachmentRecord],
-) -> Conversation:
-    """Convert records to a Conversation model.
-
-    Used by async facades and internal migration tools.
-    """
-    return Conversation.from_records(conversation, messages, attachments)
-
-
-__all__ = ["ConversationRepository", "_records_to_conversation"]
+__all__ = ["AsyncConversationRepository"]
